BOCM-358 - creates more general transitionState function
