+++ apps/api/src/server/app/case-officer/case-officer.controller.js
@@ -1,7 +1,8 @@
 import appealRepository from '../repositories/appeal.repository.js';
 import newReviewRepository from '../repositories/review-questionnaire.repository.js';
 import { reviewComplete, validateReviewRequest } from './case-officer-review.js';
-import { transitionState, lpaQuestionnaireStatesStrings } from '../state-machine/household-appeal.machine.js';
+import { lpaQuestionnaireStatesStrings } from '../state-machine/household-appeal.machine.js';
+import { transitionState } from '../state-machine/transition-state.js';
 import appealFormatter from './appeal-formatter.js';
 import CaseOfficerError from './case-officer-error.js';
 
@@ -35,7 +36,7 @@ const confirmingLPAQuestionnaire =  async function (request, response) {
 	const appeal = await getAppealForCaseOfficer(request.params.id);
 	await newReviewRepository.addReview(appeal.id, reviewResult, request.body.reason);
 	const appealStatemachineStatus = reviewResult ?  'COMPLETE' : 'INCOMPLETE';
-	const nextState = transitionState({ appealId: appeal.id }, appeal.status, appealStatemachineStatus);
+	const nextState = transitionState('household', { appealId: appeal.id }, appeal.status, appealStatemachineStatus);
 	await appealRepository.updateStatusById(appeal.id, nextState.value);
 	return response.send();
 };
