+++ apps/api/src/server/app/state-machine/household-appeal.machine.test.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-unresolved
 import test from 'ava';
 import sinon from 'sinon';
-import { transitionState } from './household-appeal.machine.js';
+import { transitionState } from './transition-state.js';
 import inspectorActionsService from './inspector.actions.js';
 import lpaQuestionnaireActions from './lpa-questionnaire.actions.js';
 
@@ -25,7 +25,7 @@ test.before('sets up mocking of actions', () => {
  */
 function applyAction(t, initial_state, action, expected_state, has_changed, context) {
 	inspectorSendBookingStub.resetHistory();
-	const next_state = transitionState(context, initial_state, action);
+	const next_state = transitionState('household', context, initial_state, action);
 	t.is(next_state.value, expected_state);
 	t.is(next_state.changed, has_changed);
 	if (next_state.value == 'awaiting_lpa_questionnaire') {
